+++ app/Http/Controllers/Admin/SrtPeDocumentsController.php
@@ -211,48 +211,60 @@ class SrtPeDocumentsController extends Controller
         $srtInputDocument = SrtInputDocument::find($request->refer_documents_id);
         //Link Media to file_upload_2
         $media = $srtInputDocument->file_upload_4->pluck('file_name')->toArray();
+        if(count($request->input('file_upload', [])) > 0){
+            // Complete Files
+            if($data['save_for'] == "Closed"){
+                $CompleteMerger = PDFMerger::init();
+                $fileUpload_1 = $srtInputDocument->getMedia('file_upload')->first();
+                $fileUpload_2 = $srtInputDocument->getMedia('file_upload_2')->first();
+                $fileUpload_3 = $srtInputDocument->getMedia('file_upload_3')->first();
+                $CompleteMerger->addPDF($fileUpload_1->getPath());
+                $CompleteMerger->addPDF($fileUpload_2->getPath());
+                $CompleteMerger->addPDF($fileUpload_3->getPath());
+                foreach ($request->input('file_upload', []) as $file) {
+                    if (count($media) === 0 || !in_array($file, $media)) {
+                        $CompleteMerger->addPDF(storage_path('tmp/uploads/' . $file), 'all');  
+                    }
+                }
+                $CompleteMerger->merge();
+                $CompleteMerger->save(storage_path('tmp/uploads/mergerPdf_Completed.pdf'), "file");
+                $srtInputDocument->addMedia(storage_path('tmp/uploads/mergerPdf_Completed.pdf'))->toMediaCollection('complete_file');
+            }
 
-         // Complete Files
-         if($data['save_for'] == "Closed"){
-            $CompleteMerger = PDFMerger::init();
-            $fileUpload_1 = $srtInputDocument->getMedia('file_upload');
-            $fileUpload_2 = $srtInputDocument->getMedia('file_upload_2');
-            $fileUpload_3 = $srtInputDocument->getMedia('file_upload_3');
-            $CompleteMerger->addPDF(public_path($fileUpload_1[0]->getUrl()));
-            $CompleteMerger->addPDF(public_path($fileUpload_2[0]->getUrl()));
-            $CompleteMerger->addPDF(public_path($fileUpload_3[0]->getUrl()));
+
+            //Merger PDF
+            $pdfMerger = PDFMerger::init();
+            //Merge PDF in Dropzone
             foreach ($request->input('file_upload', []) as $file) {
                 if (count($media) === 0 || !in_array($file, $media)) {
-                    $CompleteMerger->addPDF(storage_path('tmp/uploads/' . $file), 'all');  
+                    $pdfMerger->addPDF(storage_path('tmp/uploads/' . $file), 'all');  
                 }
             }
+            $pdfMerger->merge();
+            //Save to tmp
+            $pdfMerger->save(storage_path('tmp/uploads/mergerPdf_04.pdf'), "file");
+            
+            
+            //Delete Dropzone file on tmp 
+            foreach ($request->input('file_upload', []) as $file) {
+                if (count($media) === 0 || !in_array($file, $media)) {
+                    File::delete(storage_path('tmp/uploads/' . $file));
+                }
+            }
+            //Add media to SrtInputDocument Collection (file_upload_4)
+            $srtInputDocument->addMedia(storage_path('tmp/uploads/mergerPdf_04.pdf'))->toMediaCollection('file_upload_4');
+        }else{
+            $CompleteMerger = PDFMerger::init();
+            $fileUpload_1 = $srtInputDocument->getMedia('file_upload')->first();
+            $fileUpload_2 = $srtInputDocument->getMedia('file_upload_2')->first();
+            $fileUpload_3 = $srtInputDocument->getMedia('file_upload_3')->first();
+            $CompleteMerger->addPDF($fileUpload_1->getPath());
+            $CompleteMerger->addPDF($fileUpload_2->getPath());
+            $CompleteMerger->addPDF($fileUpload_3->getPath());
             $CompleteMerger->merge();
             $CompleteMerger->save(storage_path('tmp/uploads/mergerPdf_Completed.pdf'), "file");
             $srtInputDocument->addMedia(storage_path('tmp/uploads/mergerPdf_Completed.pdf'))->toMediaCollection('complete_file');
         }
-
-
-        //Merger PDF
-        $pdfMerger = PDFMerger::init();
-        //Merge PDF in Dropzone
-        foreach ($request->input('file_upload', []) as $file) {
-            if (count($media) === 0 || !in_array($file, $media)) {
-                $pdfMerger->addPDF(storage_path('tmp/uploads/' . $file), 'all');  
-            }
-        }
-        $pdfMerger->merge();
-        //Save to tmp
-        $pdfMerger->save(storage_path('tmp/uploads/mergerPdf_04.pdf'), "file");
-        
-        
-        //Delete Dropzone file on tmp 
-        foreach ($request->input('file_upload', []) as $file) {
-            if (count($media) === 0 || !in_array($file, $media)) {
-                File::delete(storage_path('tmp/uploads/' . $file));
-            }
-        }
-        //Add media to SrtInputDocument Collection (file_upload_4)
-        $srtInputDocument->addMedia(storage_path('tmp/uploads/mergerPdf_04.pdf'))->toMediaCollection('file_upload_4');
 
         //Delete Function on SrtInputDocument Collection (file_upload_4)
         if (count($srtInputDocument->file_upload_4) > 0) {
